@@ -6,6 +6,7 @@
     "start": "ng serve",
     "build": "ng build",
     "build:ssr": "ng run DotnetNgx:server:dev",
+    "build:ssr:watch": "npm run build:ssr -- --watch",
     "build:ssr:prod": "ng run DotnetNgx:server:production",
     "test": "ng test",
     "lint": "ng lint",
@@ -15,19 +16,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.6",
-    "@angular/common": "~7.2.6",
-    "@angular/compiler": "~7.2.6",
-    "@angular/core": "~7.2.6",
-    "@angular/forms": "~7.2.6",
-    "@angular/http": "^7.2.6",
-    "@angular/platform-browser": "~7.2.6",
-    "@angular/platform-browser-dynamic": "~7.2.6",
-    "@angular/platform-server": "7.2.6",
-    "@angular/pwa": "^0.12.4",
-    "@angular/router": "~7.2.6",
-    "@angular/service-worker": "~7.2.4",
-    "@nguniversal/module-map-ngfactory-loader": "~7.1.0",
+    "@angular/animations": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/platform-server": "8.0.0",
+    "@angular/router": "8.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "8.0.0-rc.1",
     "@ngx-translate/core": "^11.0.1",
     "@ssv/core": "^0.6.0",
     "@ssv/ngx.command": "^1.2.1",
@@ -35,30 +33,33 @@
     "@ssv/signalr-client": "^1.0.1",
     "aspnet-prerendering": "^3.0.1",
     "core-js": "^2.6.5",
-    "rxjs": "~6.4.0",
+    "rxjs": "^6.4.0",
     "tslib": "^1.9.3",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.3",
-    "@angular/cli": "~7.3.3",
-    "@angular/compiler-cli": "~7.2.6",
-    "@angular/language-service": "~7.2.6",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/cli": "8.0.6",
+    "@angular/compiler-cli": "8.0.0",
+    "@angular/language-service": "8.0.0",
     "@ssv/tools": "^0.8.1",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/jasmine": "~3.3.9",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~11.10.5",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "^4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "typescript": "3.4.5"
+  },
+  "optionalDependencies": {
+    "node-sass": "^4.9.3",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.3.3"
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1"
   }
 }
